[Refactor] #308 - zipDetail HankkiList 에서 분리
