+++ Hankkijogbo/Hankkijogbo/Global/Resources/Assets/Assets.xcassets/Button/btn_like_selected_list.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "btn_heart.svg",
+      "filename" : "ic_heart.svg",
       "idiom" : "universal"
     }
   ],
