+++ Hankkijogbo/Hankkijogbo/Present/HankkiList/View/HankkiListViewController.swift
@@ -1,126 +1,40 @@
 //
-//  ZipHeaderCollectionView.swift
+//  HankkiListViewController.swift
 //  Hankkijogbo
 //
-//  Created by 심서현 on 7/10/24.
+//  Created by 심서현 on 12/20/24.
 //
 
 import UIKit
 
-final class  HankkiListViewController: BaseViewController, NetworkResultDelegate {
+final class HankkiListViewController: BaseHankkiListViewController {
     
-    // MARK: - Properties
-    
-    private var isHeaderSetting: Bool = false
-    let type: HankkiListViewControllerType
-    let zipId: Int?
-    
-    let viewModel: HankkiListViewModel = HankkiListViewModel()
-    
-    // MARK: - UI Properties
-    
-    private let hankkiTableView = UITableView(frame: .zero, style: .grouped)
-    private lazy var emptyView = EmptyView(
-        text: type.emptyViewLabel,
-        buttonText: type == .myZip ? StringLiterals.HankkiList.moreButton : nil,
-        buttonAction: type == .myZip ? self.navigateToHomeView : nil
-    )
+    // MARK: - UI Components
     
     // MARK: - Life Cycle
     
-    init(_ type: HankkiListViewControllerType, zipId: Int?) {
-        self.type = type
-        self.zipId = zipId
-        super.init()
+    override init(_ type: BaseHankkiListViewController.HankkiListViewControllerType, zipId: Int?) {
+        super.init(type, zipId: nil)
+        self.emptyView = EmptyView(text: type.emptyViewLabel)
     }
     
-    required init?(coder: NSCoder) {
+    @MainActor required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        setupRegister()
-        setupDelegate()
-        
-        bindViewModel()
     }
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        
         setupNavigationBar()
-        if type == .myZip {
-            viewModel.getZipDetail(zipID: zipId ?? 0)
-        } else {
-            // detail같은, 이전 뷰에서 좋아요가 취소되는 경우를 고려합니다.
-            if type == .liked && !viewModel.hankkiList.isEmpty {
-                viewModel.updateMeHankkiList()
-                return
-            }
-            viewModel.getMeHankkiList(type.userTargetType)
-        }
-    }
-    
-    // MARK: - setup UI
-    
-    override func setupStyle() {
-        hankkiTableView.do {
-            $0.backgroundColor = .hankkiWhite
-            $0.isEditing = false
-            $0.bounces = true
-            $0.alwaysBounceVertical = true
-            $0.separatorStyle = .none
-        }
         
-        emptyView.do {
-            $0.isHidden = true
+        if type == .liked && !viewModel.hankkiList.isEmpty {
+            viewModel.updateMeHankkiList()
+            return
         }
-    }
-
-    override func setupHierarchy() {
-        view.addSubviews(hankkiTableView, emptyView)
-    }
-    
-    override func setupLayout() {
-        hankkiTableView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-        
-        emptyView.snp.makeConstraints {
-            $0.centerY.equalTo(type.headrViewHeight + (UIScreen.getDeviceHeight() - type.headrViewHeight) / 2)
-            $0.centerX.equalToSuperview()
-        }
+        viewModel.getMeHankkiList(type.userTargetType)
     }
 }
 
 private extension HankkiListViewController {
-    private func bindViewModel() {        
-        viewModel.reloadCollectionView = { [weak self] in
-            DispatchQueue.main.async {
-                self?.emptyView.isHidden = (self?.viewModel.hankkiList.count != 0)
-                self?.hankkiTableView.reloadData()
-            }
-            if self?.type == .myZip && !(self?.isHeaderSetting ?? false) {
-                guard let headerView = self?.hankkiTableView.headerView(forSection: 0) as? ZipHeaderTableView else { return }
-                headerView.dataBind(self?.viewModel.zipInfo ?? nil, viewModel: self!.viewModel)
-                self?.isHeaderSetting = true
-            }
-        }
-    }
-    
-    func setupRegister() {
-        hankkiTableView.register(ZipHeaderTableView.self, forHeaderFooterViewReuseIdentifier: ZipHeaderTableView.className)
-        hankkiTableView.register(ZipFooterTableView.self, forHeaderFooterViewReuseIdentifier: ZipFooterTableView.className)
-        hankkiTableView.register(HankkiListTableViewCell.self, forCellReuseIdentifier: HankkiListTableViewCell.className)
-    }
-    
-    func setupDelegate() {
-        hankkiTableView.delegate = self
-        hankkiTableView.dataSource = self
-        viewModel.delegate = self
-    }
     
     func setupNavigationBar() {
         let type: HankkiNavigationType
@@ -131,70 +45,11 @@ private extension HankkiListViewController {
             mainTitle: .string(self.type.navigationTitle),
             rightButton: .string(""),
             rightButtonAction: {},
-            backgroundColor: self.type.navigationColor
+            backgroundColor: UIColor.hankkiWhite
         )
         
         if let navigationController = navigationController as? HankkiNavigationController {
             navigationController.setupNavigationBar(forType: type)
-        }
-    }
-}
-
-extension HankkiListViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return viewModel.hankkiList.count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: HankkiListTableViewCell.className,
-                                                       for: indexPath
-        ) as? HankkiListTableViewCell else { return UITableViewCell() }
-        
-        cell.dataBind(viewModel.hankkiList[indexPath.item],
-                      isFinal: viewModel.hankkiList.count - 1 == indexPath.item,
-                      isLikeButtonDisable: self.type != .liked)
-        cell.delegate = self
-        return cell
-    }
-    
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 102
-    }
-    
-    /// 나의 족보 리스트 헤더 세팅
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        if self.type != .myZip {
-            return nil
-        }
-        
-        let headerView = tableView.dequeueReusableHeaderFooterView(
-            withIdentifier: ZipHeaderTableView.className
-        )
-        return headerView
-    }
-    
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return self.type.headrViewHeight
-    }
-    
-    /// 나의 족보 리스트 푸터 세팅
-    func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
-        if self.type != .myZip || viewModel.hankkiList.count == 0 {
-            return nil
-        }
-        
-        let footerView = tableView.dequeueReusableHeaderFooterView(
-            withIdentifier: ZipFooterTableView.className
-        ) as? ZipFooterTableView ?? ZipFooterTableView(reuseIdentifier: ZipFooterTableView.className)
-        footerView.viewController = self
-        return footerView
-    }
-    
-    func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
-        if self.type != .myZip || viewModel.hankkiList.count == 0 {
-            return 0
-        } else {
-            return 112
         }
     }
     
@@ -207,42 +62,44 @@ extension HankkiListViewController: UITableViewDataSource {
     }
 }
 
-extension HankkiListViewController: UITableViewDelegate {
+extension HankkiListViewController {
     
-    /// 나의 족보 리스트
-    /// 스와이프 해서 셀 지우기 설정
-    func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
-        if self.type != .myZip {
-            return nil
-        }
-        
-        let deleteAction = UIContextualAction(style: .destructive, title: "삭제") { (_, _, completionHandler) in
-            self.deleteItem(at: indexPath)
-            completionHandler(true)
-        }
-        deleteAction.backgroundColor = .red500
-        
-        let configuration = UISwipeActionsConfiguration(actions: [deleteAction])
-        configuration.performsFirstActionWithFullSwipe = false
-        return configuration
+    // 헤더 설정
+    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        return nil
     }
     
-    /// 나의 족보 리스트
-    /// 헤더의 스크롤 막기
-    func scrollViewDidScroll(_ scrollView: UIScrollView) {
-        if scrollView.contentOffset.y < 0 && self.type == .myZip {
-            scrollView.bounces = false
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        return 0
+    }
+    
+    // 푸터 설정
+    func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
+        return nil
+    }
+    
+    func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
+        return 0
+    }
+    
+    // cell data bind
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: HankkiListTableViewCell.className,
+                                                       for: indexPath
+        ) as? HankkiListTableViewCell else { return UITableViewCell() }
+        
+        if type == .liked {
+            cell.dataBind(viewModel.hankkiList[indexPath.item],
+                          isFinal: viewModel.hankkiList.count - 1 == indexPath.item,
+                          isLikeButtonDisable: false)
+            cell.delegate = self
+            
+            return cell
         } else {
-            scrollView.bounces = true
+            cell.dataBind(viewModel.hankkiList[indexPath.item],
+                          isFinal: viewModel.hankkiList.count - 1 == indexPath.item)
+            return cell
         }
-    }
-    
-    /// 터치시 식당 디테일로 이동
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        tableView.deselectRow(at: indexPath, animated: true)
-    
-        let hankkiDetailViewController = HankkiDetailViewController(hankkiId: viewModel.hankkiList[indexPath.item].id)
-        navigationController?.pushViewController(hankkiDetailViewController, animated: true)
     }
 }
 
@@ -273,39 +130,5 @@ extension HankkiListViewController: HankkiListTableViewCellDelegate {
             
             viewModel.hankkiList[indexPath.item] = newData
         }
-    }
-}
-
-private extension HankkiListViewController {
-    /// 셀을 지우는 함수
-    func deleteItem(at indexPath: IndexPath) {
-        let request: DeleteZipToHankkiRequestDTO = DeleteZipToHankkiRequestDTO(favoriteId: zipId ?? 0, storeId: viewModel.hankkiList[indexPath.item].id)
-        viewModel.deleteZipToHankki(requestBody: request) {
-            self.removeCellFromCollectionView(indexPath)
-        }
-    }
-    
-    private func removeCellFromCollectionView(_ indexPath: IndexPath) {
-        viewModel.hankkiList.remove(at: indexPath.row)
-        
-        hankkiTableView.beginUpdates()
-        hankkiTableView.deleteRows(at: [indexPath], with: .automatic)
-        hankkiTableView.endUpdates()
-    }
-}
-
-extension HankkiListViewController {
-    /// Home View로 이동하는 함수
-    func navigateToHomeView() {
-        var rootViewController = self.view.window?.rootViewController
-
-        if let navigationController = rootViewController as? UINavigationController {
-            rootViewController = navigationController.viewControllers.first
-        }
-        if let tabBarController = rootViewController as? UITabBarController {
-            tabBarController.selectedIndex = 0
-        }
-        
-        navigationController?.popToRootViewController(animated: true)
     }
 }
