+++ Hankkijogbo/Hankkijogbo/Present/Home/View/HomeViewController.swift
@@ -194,7 +194,7 @@ extension HomeViewController {
         if let zipId = notification.userInfo?["zipId"] as? Int {
             
             self.showBlackToast(message: StringLiterals.Toast.addToMyZipBlack) { [self] in
-                let hankkiListViewController = HankkiListViewController(.myZip, zipId: zipId)
+                let hankkiListViewController = ZipDetailViewController(zipID: zipId)
                 navigationController?.pushViewController(hankkiListViewController, animated: true)
             }
         }
