+++ Hankkijogbo/Hankkijogbo/Present/Report/Complete/ReportCompleteViewController.swift
@@ -252,7 +252,7 @@ private extension ReportCompleteViewController {
         if let zipId = notification.userInfo?["zipId"] as? Int {
             
             self.showWhiteToast(message: StringLiterals.Toast.addToMyZipWhite) { [self] in
-                let hankkiListViewController = HankkiListViewController(.myZip, zipId: zipId)
+                let hankkiListViewController = ZipDetailViewController(zipID: zipId)
                 navigationController?.pushViewController(hankkiListViewController, animated: true)
             }
         }
