+++ Hankkijogbo/Hankkijogbo/Present/ZipList/View/ZipListViewController.swift
@@ -130,7 +130,7 @@ private extension ZipListViewController {
     }
     
     func navigateToHankkiListViewController(zipId: Int) {
-        let hankkiListViewController = HankkiListViewController(.myZip, zipId: zipId)
+        let hankkiListViewController = ZipDetailViewController(zipID: zipId)
             navigationController?.pushViewController(hankkiListViewController, animated: true)
     }
